Creación controlador producto

+++ src/main/java/com/curso/SpringBoot/persistence/crud/ProductoCrudRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface ProductoCrudRepository extends CrudRepository<Producto, Long> {
+public interface ProductoCrudRepository extends CrudRepository<Producto, Integer> {
     //Existe dos formas, llamando nativamente o con los query methodsyp
   //  @Query(value = "SELECT * FROM productos WHERE id_categoria = ?", nativeQuery = true)
     List<Producto> findByIdCategoriaOrderByNombreAsc(int idCategoria);
